refactor(tools): restructure tool dir
