+++ api/app/langchain/tools/index.js
@@ -1,10 +1,19 @@
+const GoogleSearchAPI = require('./GoogleSearch');
+const HttpRequestTool = require('./HttpRequestTool');
+const AIPluginTool = require('./AIPluginTool');
+const OpenAICreateImage = require('./DALL-E');
+const StructuredSD = require('./structured/StableDiffusion');
+const StableDiffusionAPI = require('./StableDiffusion');
+const WolframAlphaAPI = require('./Wolfram');
 const SelfReflectionTool = require('./SelfReflection');
-const availableTools = require('./manifest.json');
-const { validateTools, loadTools } = require('./handleTools');
 
 module.exports = {
-  validateTools,
-  loadTools,
-  availableTools,
+  GoogleSearchAPI,
+  HttpRequestTool,
+  AIPluginTool,
+  OpenAICreateImage,
+  StructuredSD,
+  StableDiffusionAPI,
+  WolframAlphaAPI,
   SelfReflectionTool
-};
+}
