+++ api/app/langchain/tools/util/index.js
@@ -1,0 +1,8 @@
+const availableTools = require('../manifest.json');
+const { validateTools, loadTools } = require('./handleTools');
+
+module.exports = {
+  validateTools,
+  loadTools,
+  availableTools
+};
